rft: Move build info to pkg
