+++ pkg/server/web/beego.go
@@ -22,9 +22,6 @@ import (
 )
 
 const (
-	// VERSION represent beego web framework version.
-	VERSION = "1.12.2"
-
 	// DEV is for develop
 	DEV = "dev"
 	// PROD is for production
